internal/v1: split out identity tests
