+++ internal/v1/server_test.go
@@ -202,12 +202,6 @@ func TestWithoutOsbuildComposerBackend(t *testing.T) {
 	}()
 	defer tokenSrv.Close()
 
-	t.Run("VerifyIdentityHeaderMissing", func(t *testing.T) {
-		respStatusCode, body := tutils.GetResponseBody(t, "http://localhost:8086/api/image-builder/v1/version", nil)
-		require.Equal(t, 400, respStatusCode)
-		require.Contains(t, body, "missing x-rh-identity header")
-	})
-
 	t.Run("GetVersion", func(t *testing.T) {
 		respStatusCode, body := tutils.GetResponseBody(t, "http://localhost:8086/api/image-builder/v1/version", &tutils.AuthString0)
 		require.Equal(t, 200, respStatusCode)
@@ -333,70 +327,8 @@ func TestWithoutOsbuildComposerBackend(t *testing.T) {
 		}
 	})
 
-	t.Run("AccountNumberFallback", func(t *testing.T) {
-		respStatusCode, _ := tutils.GetResponseBody(t, "http://localhost:8086/api/image-builder/v1/version", &tutils.AuthString0WithoutEntitlements)
-		require.Equal(t, 200, respStatusCode)
-	})
-
-	t.Run("BogusAuthString", func(t *testing.T) {
-		auth := "notbase64"
-		respStatusCode, body := tutils.GetResponseBody(t, "http://localhost:8086/api/image-builder/v1/version", &auth)
-		require.Equal(t, 400, respStatusCode)
-		require.Contains(t, body, "unable to b64 decode x-rh-identity header")
-	})
-
-	t.Run("BogusBase64AuthString", func(t *testing.T) {
-		auth := "dGhpcyBpcyBkZWZpbml0ZWx5IG5vdCBqc29uCg=="
-		respStatusCode, body := tutils.GetResponseBody(t, "http://localhost:8086/api/image-builder/v1/version", &auth)
-		require.Equal(t, 400, respStatusCode)
-		require.Contains(t, body, "does not contain valid JSON")
-	})
-
-	t.Run("EmptyAccountNumber", func(t *testing.T) {
-		// AccoundNumber equals ""
-		auth := tutils.GetCompleteBase64Header("000000")
-		respStatusCode, _ := tutils.GetResponseBody(t, "http://localhost:8086/api/image-builder/v1/version", &auth)
-		require.Equal(t, 200, respStatusCode)
-	})
-
-	t.Run("EmptyOrgID", func(t *testing.T) {
-		// OrgID equals ""
-		auth := tutils.GetCompleteBase64Header("")
-		respStatusCode, body := tutils.GetResponseBody(t, "http://localhost:8086/api/image-builder/v1/version", &auth)
-		require.Equal(t, 400, respStatusCode)
-		require.Contains(t, body, "invalid or missing org_id")
-	})
-
 	t.Run("StatusCheck", func(t *testing.T) {
 		respStatusCode, _ := tutils.GetResponseBody(t, "http://localhost:8086/status", nil)
-		require.Equal(t, 200, respStatusCode)
-	})
-}
-
-func TestOrgIdWildcard(t *testing.T) {
-	srv, tokenSrv := startServer(t, "", "")
-	defer func() {
-		err := srv.Shutdown(context.Background())
-		require.NoError(t, err)
-	}()
-	defer tokenSrv.Close()
-
-	t.Run("Authorized", func(t *testing.T) {
-		respStatusCode, _ := tutils.GetResponseBody(t, "http://localhost:8086/api/image-builder/v1/version", &tutils.AuthString0)
-		require.Equal(t, 200, respStatusCode)
-	})
-}
-
-func TestAccountNumberWildcard(t *testing.T) {
-	srv, tokenSrv := startServer(t, "", "")
-	defer func() {
-		err := srv.Shutdown(context.Background())
-		require.NoError(t, err)
-	}()
-	defer tokenSrv.Close()
-
-	t.Run("Authorized", func(t *testing.T) {
-		respStatusCode, _ := tutils.GetResponseBody(t, "http://localhost:8086/api/image-builder/v1/version", &tutils.AuthString0)
 		require.Equal(t, 200, respStatusCode)
 	})
 }
